+++ frontend/src/components/button/submitButton.css
@@ -1,38 +1,38 @@
 .submit__button {
-    box-shadow: 0rem 0rem 0rem 0rem #111111;
-    @apply w-full;
+  box-shadow: 0rem 0rem 0rem 0rem #111111;
+  @apply w-full;
 }
 
 .submit__button:hover {
-    animation: button__animation--mouseover 0.15s ease-in-out forwards;
+  animation: button__animation--mouseover 0.15s ease-in-out forwards;
 }
 
 .submit__button--animation {
-    animation: button__animation--mouseleave 0.15s ease-in-out forwards;
+  animation: button__animation--mouseleave 0.15s ease-in-out forwards;
 }
 
 @keyframes button__animation--mouseover {
-    0% {
-        transform: translate(0, 0);
-        box-shadow: 0rem 0rem 0rem 0rem #111111;
-    }
-    100% {
-        transform: translate(-0.125rem, -0.125rem);
-        box-shadow: 0.25rem 0.25rem 0rem 0rem #111111;
-    }
+  0% {
+    transform: translate(0, 0);
+    box-shadow: 0rem 0rem 0rem 0rem #111111;
+  }
+  100% {
+    transform: translate(-0.125rem, -0.125rem);
+    box-shadow: 0.25rem 0.25rem 0rem 0rem #111111;
+  }
 }
 
 @keyframes button__animation--mouseleave {
-    0% {
-        transform: translate(-0.125rem, -0.125rem);
-        box-shadow: 0.25rem 0.25rem 0rem 0rem #111111;
-    }
-    100% {
-        transform: translate(0, 0);
-        box-shadow: 0rem 0rem 0rem 0rem #111111;
-    }
+  0% {
+    transform: translate(-0.125rem, -0.125rem);
+    box-shadow: 0.25rem 0.25rem 0rem 0rem #111111;
+  }
+  100% {
+    transform: translate(0, 0);
+    box-shadow: 0rem 0rem 0rem 0rem #111111;
+  }
 }
 
 .submit__button__content {
-    @apply flex items-center justify-center pt-[1rem] pb-[1rem] border-sm border-text-primary box-border bg-point-lavender;
+  @apply flex items-center justify-center pt-[1rem] pb-[1rem] border-sm border-text-primary box-border bg-point-lavender;
 }
