+++ frontend/src/components/button/submitButton.tsx
@@ -1,37 +1,40 @@
 import { useRef } from 'react'
+import Text from '../text/text'
 
 interface SubmitButtonProps {
-    text: string
-    width?: number
-    height?: number
-    handleOnClick?: () => void
+  text: string
+  width?: number
+  height?: number
+  handleOnClick?: () => void
 }
 
 /**
  * @param text 버튼의 텍스트
  */
 const SubmitButton = ({ text, handleOnClick }: SubmitButtonProps) => {
-    const submitButtonRef = useRef<HTMLButtonElement>(null)
+  const submitButtonRef = useRef<HTMLButtonElement>(null)
 
-    const handleMouseOver = () => {
-        if (submitButtonRef.current) {
-            submitButtonRef.current.classList.add('submit__button--animation')
-        }
+  const handleMouseOver = () => {
+    if (submitButtonRef.current) {
+      submitButtonRef.current.classList.add('submit__button--animation')
     }
+  }
 
-    return (
-        <button
-            type="submit"
-            className="submit__button"
-            ref={submitButtonRef}
-            onClick={handleOnClick}
-            onMouseOver={handleMouseOver}
-        >
-            <div className="submit__button__content">
-                <span className="display-regular-14">{text}</span>
-            </div>
-        </button>
-    )
+  return (
+    <button
+      type="submit"
+      className="submit__button"
+      ref={submitButtonRef}
+      onClick={handleOnClick}
+      onMouseOver={handleMouseOver}
+    >
+      <div className="submit__button__content">
+        <Text size={14} color="text-primary">
+          {text}
+        </Text>
+      </div>
+    </button>
+  )
 }
 
 export default SubmitButton
