+++ frontend/src/components/input/input.css
@@ -1,33 +1,33 @@
 .input__wrapper {
-    @apply w-full flex flex-col;
+  @apply w-full flex flex-col;
 }
 
 .input__wrapper:focus-within .input__border {
-    @apply border-b-sm border-point-blue;
+  @apply border-b-sm border-point-blue;
 }
 
 .input__icon_wrapper {
-    @apply flex flex-row items-center;
+  @apply flex flex-row items-center;
 }
 
 .input {
-    background-color: transparent;
-    @apply flex-1 p-sm;
+  background-color: transparent;
+  @apply flex-1 p-sm;
 }
 
 .input[type='text'] {
-    @apply display-regular-14;
+  @apply display-regular-14;
 }
 
 .input::placeholder {
-    @apply display-regular-14;
+  @apply display-regular-14;
 }
 
 .input:focus {
-    @apply outline-none;
+  @apply outline-none;
 }
 
 .input__border {
-    transition: all 0.2s ease-in-out;
-    @apply border-b-sm border-text-primary;
+  transition: all 0.2s ease-in-out;
+  @apply border-b-sm border-text-primary;
 }
