--- frontend/src/components/menu/menu.css
@@ -1,29 +1,0 @@
-.menu__wrapper {
-    @apply w-full flex flex-col items-center border-t-sm border-contour-primary;
-}
-
-.menu__animated__bar__wrapper {
-    @apply w-[15rem];
-}
-
-.menu__animated__bar {
-    transition: all 0.2s ease-in-out;
-    @apply m-[-0.065rem] w-[5rem] h-[0.125rem] bg-text-primary;
-}
-
-.menu__button {
-    transition: all 0.2s ease-in-out;
-    @apply w-[5rem];
-}
-
-.menu__button span {
-    @apply display-regular-12;
-}
-
-.menu__button--selected {
-    @apply text-text-primary;
-}
-
-.menu__button--unselected {
-    @apply text-text-secondary;
-}
